+++ CHANGELOG.md
@@ -10,6 +10,8 @@ All notable changes to this project will be documented in this file.
  * Fixed In Scope on sequence diagram
  * Fixed UML Primitive behavior when resizing
  * Added Async Message link element to Sequence Diagram
+ * Added Flat Message link element to Sequence Diagram
+ * Added Return Message link element to Sequence Diagram
 
 ## [1.1.2-beta.3] - 2025-03-10
 ### Fixed